zdtm: use architecture-specific atomic operations
The test static/session01 fails to compile on ARM
because it uses x86-specific atomic operations.

It seems unfeasible to implement architecture-specific
operations for the test-suite so the build system is modified
to use the same atomic operation implementation both for
crtools and its test suite.

The header test/zdtm/lib/lock.h is modified to use
other declaration of the type atomic_t.